Revert "DXP-1310 Perform StreamX availability check before indexing data (#15)"
This reverts commit 54bf0f3893b5583b31225c49182ae20b380fc6d5.
