+++ composer.json
@@ -20,7 +20,7 @@
   ],
   "require": {
     "php": ">=7.4",
-    "streamx/ingestion-client": "1.0.1",
+    "streamx/ingestion-client": "1.0.0",
     "magento/framework": ">=101.0.0",
     "magento/module-store": ">=100.2.0",
     "magento/module-backend": ">=100.2.0",
