+++ src/catalog/test/integration/AppEntityUpdateStreamxPublishTests/BaseAppEntityUpdateTest.php
@@ -2,13 +2,13 @@
 
 namespace StreamX\ConnectorCatalog\test\integration\AppEntityUpdateStreamxPublishTests;
 
-use StreamX\ConnectorCatalog\test\integration\BaseStreamxConnectorPublishTest;
+use StreamX\ConnectorCatalog\test\integration\BaseStreamxPublishTest;
 use StreamX\ConnectorCatalog\test\integration\utils\MagentoIndexerOperationsExecutor;
 
 /**
  * @inheritdoc
  */
-abstract class BaseAppEntityUpdateTest extends BaseStreamxConnectorPublishTest {
+abstract class BaseAppEntityUpdateTest extends BaseStreamxPublishTest {
 
     protected abstract function indexerName(): string;
