+++ src/catalog/test/integration/DirectDbEntityUpdateStreamxPublishTests/BaseDirectDbEntityUpdateTest.php
@@ -2,13 +2,13 @@
 
 namespace StreamX\ConnectorCatalog\test\integration\DirectDbEntityUpdateStreamxPublishTests;
 
-use StreamX\ConnectorCatalog\test\integration\BaseStreamxConnectorPublishTest;
+use StreamX\ConnectorCatalog\test\integration\BaseStreamxPublishTest;
 use StreamX\ConnectorCatalog\test\integration\utils\MagentoIndexerOperationsExecutor;
 
 /**
  * @inheritdoc
  */
-abstract class BaseDirectDbEntityUpdateTest extends BaseStreamxConnectorPublishTest {
+abstract class BaseDirectDbEntityUpdateTest extends BaseStreamxPublishTest {
 
     protected abstract function indexerName(): string;
